+++ globals.h
@@ -7,4 +7,4 @@ Arduboy2 arduboy;
 ArduboyTones sound(arduboy.audio.enabled);
 Sprites sprites;
 
-byte ground = HEIGHT - 10;
+int ground = HEIGHT - 10;
